requirements.txt
fastapi>=0.110.0
uvicorn>=0.27.0
langchain>=0.1.0
langchain-aws>=0.1.0
boto3>=1.34.0
tenacity>=8.2.0
python-dotenv>=1.0.0
pydantic>=2.6.0
pydantic-settings>=2.0.0
requests>=2.31.0
python-multipart>=0.0.9
grpcio>=1.60.0
grpcio-tools>=1.60.0
grpcio-health-checking>=1.60.0
beautifulsoup4>=4.12.0
chromadb>=0.4.0
langchain-community>=0.0.10
langchain-text-splitters>=0.0.1
psycopg2-binary>=2.9.0
openai>=1.0.0
redis==4.6.0 # Downgraded for LangChain Compatibility
sqlalchemy>=2.0.0
asyncpg>=0.29.0
pytz
influxdb-client>=1.36.0
google-api-python-client>=2.0.0
google-auth>=2.0.0
google-auth-oauthlib>=1.0.0
aiokafka>=0.10.0
pgvector
pycryptodome>=0.2.0
duckduckgo-search>=5.0.0
grpcio-health-checking>=1.60.0
pycryptodome>=3.20.0
groq>=0.4.0
